pyproject.toml
[tool.black]
line-length = 80

[tool.poetry]
name = "radio_core"
version = "0.1.0"
description = ""
authors = ["Sergey Sharybin <sergey.vfx@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9,<3.13"

[tool.poetry.group.dev.dependencies]
numpy = "^1.24.1"
pillow = "^9.4.0"
matplotlib = "^3.6.3"
scipy = "1.11.1"
